requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
tensorflow
tensorboard
openpyxl
nltk
flask
streamlit
pymongo[srv]==3.12
python-dotenv
tqdm